Don't update cursor position when the last pointer goes up
In the case of a two-finger right click where the first finger goes
up before the second one, ensure that we don't click at the first finger
and then move the cursor to the second finger.  If we have a genuine
drag, there will be at least one ACTION_MOVE event, and the position
will be updated then.